Refactor auth module to use requests auth plugins
This also upgrades the requests version to the latest
(1.1.0).

One of the problems with the request auth plugins is
that the auth handlers are invoked on a more or less
prepared requests.  This means the query params have
been encoded and the post body params have also been
urlencoded.  This means that we have to re-parse the
query string and post body params, process/modify them
as needed, the re-urlencode them.  Not ideal.

Sigv2 auth handler implementation

Update sigv4 to use requests auth plugin