Fix s3 auth
There were two problems from what I could tell.
One is that iterating over headers was broken in
python2 because the __iter__ method on the Message
class is py3 only.

The second was that the hex signature was a bytes
array and not str, which in python3 was being sent
as "b'signature'".

Verified that unit tests pass, and that I can make
s3 calls via the aws-cli in both py2 and py3.