Add py_input_token when building pagination configs
The operation.call method takes the kwargs as the python
version of the arg name (e.g. next_token not NextToken).
Rather than doing this at runtime, I added a py_input_token to
the model that is added during build time.

If we need to manually override this process, we can just add
py_input_token to the .extra.json files ourselves, and they will
be used instead.