Add table module to generate multitables
These are tables with multiple sections.  Each section
has a title, a row of headers, and rows of data.  A
section does not need a title or headers.  The table can also
generates output with colors, though there are options to
turn this off.

This pulls in a dependency on blessings for terminal output
as well as terminal width detection.