Add automatic pagination support to CLI
I've added a TODO that we need to http responses for paginated
output but I think the plugin refactoring will address that.

The interface is written in such a way to allow for streaming
formatter responses, but the built in formatters all buffer their
objects. Eventually we should add streaming output formatters but
for formats like json we'll need a streaming json parser.