Change profile section names in aws config file
Profile names are now specified as::

    [profile "name"]
    aws_... = foo

The motivation for this change is to "make room" for other
configuration options in the config file.  Currently, any section
name in the config file is considered a potential profile name.

Another alternative is to just whitelist a set of configuration
options that aren't profiles, but then that restricts what names
you can use for profiles.