Change file prefix from file: to file://
Exact same semantics as before except the prefix has to be file://.
This was done to prevent a conflict when using the shorthand notation
for complex param (we wouldn't be able to tell the difference between
the file: prefix and a key named 'file' or 'http' of 'https').

Docs updated as well.