Update search path processing to work on windows
We have to use os.path.abspath because this will capture the
current drive we're on.  Otherwise if our cwd is on a different
drive, we'll build our search path based on the current drive.

I've also switched the path separator to os.pathsep.

See (aws/aws-cli#101) for more details.