Don't add None responses to multi result key pagination
Also, switch from defaultdict to just creating the keys
with empty lists up front.  This is because now, even if one
of the result keys is always None, we still should have a top level
result key that's just an empty list, which wouldn't happen if we
were using a defaultdict.