Return value from help command
Even though this command will call sys.exit, we still
want it to return, if nothing else for clarity.

This also helps for testing purposes as well.

I added a start of tests for docs to make sure that we don't
break help output in the future.  I plan on greatly expanding these
tests as needed.