Fix streaming output arg customization
I noticed that there actually are unit tests for this, but
we weren't checking the RC of the run command, only that the request
params looked how we wanted them to.  As a result, I've updated the
base test class to check for expected RCs from the commands that are
run.  This caused the tests to fail and should ensure that if we
break the streaming output arg customization, we'll get a test failure.