Change event name of top command table
Now that the root command table is 'building-command-table'
and service subcommands are 'building-command-table.servicename',
this makes it hard for a handler that wants to add a new top level
command as they have to parse the event name.  In this change,
the handlers can subscribe to 'building-command-table.main' which
makes it easy to subscribe to the event you want.