Added cp command and improved ls command.
The cp command overrides get, copy, and
put commands by relying on positional arguments.
Functionality was expanded for mv so that it
can move files locally to s3 and vice versa.
The ls command was improved so it does not
require a path argument to list all S3 buckets.
All tests were changed to account for change in
commands.