Don't type convert to a list
While this is useful in some cases, other times this masks
hard to find issues.

If a user provides the wrong value for a list param, we will
automatically convert the value to a [value].  This is convenient,
provided that the user wanted to do this, but it also makes it harder
to track down errors.  Now you'll get an error if the object
is not a list.