Fixed a few bugs and added more abstraction.
As to bugs, reenabled ability to cleanup of multipart
uploads when the operations are interrupted (i.e. Ctrl-C).
Also, ensure files that are larger to 5 TB cannot be uploaded
to S3.

As to abstraction, methods to the Executer class were implemented
such as submit() and wait() in order to take the handling of the main
queue out of task.py and fileinfo.py and into the Executer class.
This will make it easier to change how to handle tasks that are
submitted if the threading strategy ever moves away from the use
of a main queue.