Complain when file/http param resource does not exist
If a user provides an arg via file://, http://, https://
and that resource does not exist, we should immediately stop
and let them know they're specifying a resource that does not
exist.  Otherwise the user will generally get an error message
further downstream that makes it harder to troubleshoot.  Since
most of these file params are used for complex types, this generally
happens in the json parsing and/or the arg shorthand processing.