Fix html rendering docs
So CLICommand have grown a create_help_command which looks like
it was added with the doc changes a while back.

Also htmlgen is reaching into the service_command for a
'_service_object' which needs to be fixed in a more long term way,
but for the time being I just added a hasattr check so we can
render docs.