Use the command table names for operation help
This repurposes the event_name attr to be the
name from the command table.  The examples customization
has also been updated to use the event_class.  This means
that the s3api examples can be placed in the examples/s3api
folder.

I've added the name attr on the command objects as part of the
public interface that all subclasses must implement.