Relax queue timeouts
This brings the idling on task queues from ~30%
to about 1-2%.  Note that this does influence the
responsiveness of the UI, particularly with interruption
and executor shutdown (from .1 to 1 second).  I think
this will be addressed when reworking the executor code.

One of the fixes for #333, but note that this is not the
main contribution of high CPU utilization so it does not
completely fix the issue, this will require changing the
multipart upload task algorithms.