Added --compare-on-etag option for s3 sync
This option compares the contents of buckets and local directories based
on the ETag of the S3 objects or md5 of the local files. This is useful
for scenarios in which other processes may be updating objects in the
target bucket, updating the timestamps and changing the contents, but
keeping the size of the objects the same.

This required many modifications to FileGenerator to make the etag
available to the FileInfo class.

Currently passes all unit tests, failing some integration tests.