Add support for a "configure set" command
This allows a user to set a single configuration variable
without having to edit the contents of the config file.

This makes it possible to completely shield the user from knowing
where the config file is.

This supports the same syntax as "get" with respect to the qualified
and unqualified names.