Move uri param processing into an event handler
Previously this logic was baked into the argument
subclass that mapped to parameters in the JSON model.
However, non model arugments should be able to leverage this ability
without subclassing the JSON model specific argument class.

The change here is to move that processing into an event.
There's a `load-cli-arg` event that fires.  The `get_paramfile`
hooks into this and performs its logic.