Remove session from paramfile interface
The session variable isn't being used and makes it an unnecessary
requirement for code that wants to use this functionality.

The session variable would be useful for code that needs to make
AWS requests.  For example, if we support `s3://` prefix in the future
(there's an existing issue for this) then *that* code would need to have
a reference to a session.  However, I think each handler should be
responsible for getting its necessary dependencies.