Fix issue with shadowed max-items arguments
When we add pagination configs for a service, we hide
the original manual pagination arguments in favor of the normalized
arguments.  However, to support backwards compatability we still
allow users to specify the manual arguments.  If we detect this is the
case we will turn off pagination.

However, when the normalized argument matches the manual paging argument
(as is the case when a service users `MaxItems`), we have an ambiguous
case.  Did they mean the back-compat version of the normalized argument
version?  Previously we were assuming the user meant the manual paging
argument.  Now we assume that they meant the normalized paging
arguments.

Services impacted:

* route53
* iam
* ses