Don't validate --endpoint-url for help subcommand
"aws cloudsearchdomain" has --endpoint-url validation because
it's a required option for that service.  However, we should
*not* be validating this when a user is just trying to get
help.

The fix is to move the validation to the operation-args-parsed.*
event, which is after all the help processing code.

I've also added integration tests that actually run the help command
for cloudsearchdomain to verify this behavior works.