EMR: Add validation logic in steputils and clean up create-cluster and add-steps examples.
This fix adds a validation on customer provided list, which checks if the list is empty.
For example, if customers create a cluster with a command like
"aws emr create-cluster --ami-version 3.1.0  --instance-type m1.large --steps Type=Pig,Args=",
the commmand line will give the error message that "aws: error: The prameter Args cannot be an empty list.".

This commit removes Hive and Pig "Version" from create-cluster and add-steps examples.