Only display --page-size when operation is paginated
This moves the --page-size CLI option to the pagination customization.
Before adding the argument, we first verify that not only can the
operation be paginated, but also that the pagination configuration
supports a limit key.  This also means that the --page-size argument
is now documented on the man/html page for each operation that supports
pagination, making it more clear when this argument can be used.

The only other update was that I had to add a page size to the
S3 plugin as well.