Add customization to alter scalar parsing behavior
This fixes #1001.  This also fixes a regression we had
when displaying binary content.

As mentioned on the github issue, we were previously always
showing binary content as base64 encoded.  This PR adds back
this behavior.

In addition, this also gives us a hook into changing the
timestamp parsing so we can later add a config option to
unify the way timestamps are displayed.

Also fixes #970, and closes #1005.

I've also added an integration test for #1001 to ensure
we are in fact displaying a response that can be printed to
stdout.