Remove session dep from CommandParameters
CommandParameters is moving (but is not quite there) towards being
only concerned with command parameter validation and restructuring.
It's should not be concerned with also making S3 requests, _especially_
recursively deleting customer objects.  This is better suited for
the CommandArchitecture or the RbCommand objects.  While it would
take more work to get it there, the S3TransferCommand base class seems
to be a better fit and an incremental improvement.  It makes the
CommandParameters class simpler.