Convert cloudtrail cusotmization to clients
Also added a bit to the test suite as there some components
I was changing that weren't tested.

The only weird thing I had to add was unregistering the error handler
the awscli uses.  Botocore clients raise exceptions now, but with this
customization in place the AWS CLI's client error was being raised
instead of botocores, which is what we want.