Fix issue when disabling pagination
Fixes #1247.

This was a bug introduced in the botocore client switchover.
The CLI has a feature where, for backcompat, it will detect if the
user is specifying manual pagination args and implicitly turn off
pagination.

This functionality was broken when we shadow an operation argument
with one of the pagination args we add.  Before, this was silently
working because, in this scenario, "max_items" is a valid argument
to both ".paginate()" and "operation.call()".

With the client switchover we need to be more explicit.  If we intend
for "max_items" to go to the paginator it *must* be "max_items" (note
the snake_case).  If it's intended for the operation it *must* be
in the casing originally specified by the service (in this case
it was "MaxItems").

There's not really a clean way to fix this, so what I ended up doing
was tracking any shadowed args, and in the case where we need to turn
off pagination, we put back the arg objects we shadowed.