Swap out old shorthand parser with new parser
I needed to update a lambda test because it was assuming
that the shorthand parser would throw an error if you specify a
key that's not in the model.  Now that this is no longer the case,
this needs to be handled in the customization itself.

The one remaining task is to handle documentation for shorthand.
For now, the existing docs are left in place.