Fix a bug when creating emr role with AWS_CA_BUNDLE
It is reported that there is a bug causing failure to create emr role,
when customer has an AWS_CA_BUNDLE setting in the environment.

When the bug occurs, running::

    aws emr create-default-roles

fails with "SSLError" object has no attribute.

This commit should be able to fix that, and also fix the incorrect test cases.