Catch more specific exceptions in assume role provider
To repro, call an aws command with an explicit ``--region`` and
a non-existent profile:

  aws ec2 describe-instances --region us-east-1 --profile does-not-exist

The issue was that the assume role provider was catching too general
of an exception so the profile does not exist error was not being
propogated.

There's also a bug in botocore with how we're handling failures
for lazily registered components.  This is why we're seeing the
"Unknown component" bug.  This will require a separate PR
to botocore.