Catch a bug in mimetype library
Issue was that on windows some MIME types in the Windows library could not
be encoded using the system default encoding. This has been fixed in later
versions of python, but we should not hard fail if the user is using a python
version without the fix. Instead just catch the exception and not set the
content type.