Add some path checks around user provided paths
This mainly applies to the existance of user provided local paths.
If the transfer is an upload and that local path does not exist, this will
cause the command to fail instead of skipping. If the transfer is a
syncing download and the local path does not existant, there should not
be warnings.