Skip known glacier incompatible operations
Before we would try to process an operation (FileInfo) no matter what it was
if it reached the S3Handler. Now we do some checks to see if the operation
will fail given we know the FileInfo object represents a Glacier objects. If
it is we know skip it as opposed to letting the error handling handle it
which is much slower if you have large/many Glacier objects.