Add backcompat args for argument renames
In boto/botocore#703, there were some correction made
to the CamelCase->snake_case conversions.

However, in the CLI, we still need to support these arguments.

This fix for this is to:

* Add a copy of the correct arg into the arg table with the
  "wrong" name.
* Mark the "wrong" name entry as undocumented.

There were no unit tests for this module, so I added a few.
I also added some functional tests for some (but not all)
the rename cases, just to ensure the end to end functionality
is working as we expect.