Refactor FileGenerator call()
The list_files and list_objects no longer yield the same data. The data
required from s3 was no longer the same as the data required from a local
file system. The data yielded back from both methods was minimized such
that a minimized number of variables needed to be yielded.