Ignore Ctrl-C when paging help output
Simply catching the Ctrl-C will typically leave
the user's terminal in a bad state that will
require a ``reset`` to fix.

This new behavior is also consistent with
the default pager of less as well as reading
man pages via ``man``.  If you try something like
``man rsync`` you can't Ctrl-C out of that.  You have
to use ``q`` to exit.