Switch test code over to skip_if_windows
This fixes an issue with test_cli, where a test needed
to be annotated to be skipped on windows but was missed.

I've also added a test utility function for this and I've
updated all the existing skippable tests to use this decorator.
This also makes is easier in the future if we need to update
what qualifies as "not windows."