Take into account if a glacier object is restored for s3 downloads
If a glacier object has been restored, we should be able
to download the file.

Fixes a regression introduced when we started ignoring the
GLACIER storage class.

This doesn't fix the issue for a recursive downloads/sync,
but that isn't possible because the Restore data isn't
in a ListObjects request.