Ignore ProfileNotFound in assumerole cache injector
If a user has provided a profile that does not exist, trying to retrieve
components/config on the session will raise ProfileNotFound.  Sometimes
this is invalid:

"ec2 describe-instances --profile unknown"

and sometimes this is perfectly valid:

"configure set region us-west-2 --profile brand-new-profile"

Because we can't know (and don't want to know) whether the customer is
trying to do something valid, we just immediately return.  If it's
invalid something else up the stack will raise ProfileNotFound,
otherwise the configure (and other) commands will work as expected.

The previous behavior had a blanket "except Exception", but
I don't think this is necessary.

I also updated the tests to have better error messages.  They
were previously failing with "filename <foo> not found".