Use the list index as key name when visiting list types
In the backwards compatible shorthand visitor, we need to
pass the list index as the key into the parent structure
when we visit sub elements in the list.  This is necessary
in the case where we need to modify a child element and
reinsert it back into the parent structure.