Mark args as not required when aliasing
If you have a required argument that you'd like to rename,
we were previously copying the argument and injecting
the copy into the arg table under a different name.

The problem is that the ``required`` field is also copied.
This will require the user to provide both arguments, which
defeats the point of aliasing the argument we want to hide.

The change here is to not make aliased args required.  That
way the user can specify either version.