Update ZipFile parameter documentation to be consistent.
When we pull ZipFile out of the Code argument in CreateFunction we
give it a docstring relevant to the CLI, but we don't give that
doc string in UpdateFunctionCode. This causes confusion for the
customer because the service documentation mentions you have to
base64 encode the .zip, which we handle for the customer.

Closes #1704

Support stubbing out operations.