Gracefully handle encoding errors when printing s3 keys
When displaying S3 key names via the high level
``s3`` commands, it's possible to encounter an S3 key that
cannot be encoded using the user's locale settings.

For example, if I upload a key on linux that contains
non-ascii characters, I'll upload the key to s3 with
utf-8 encoding.

If I then try to ``aws s3 ls`` on windows (which for example
might be using an active code page of 437), then the
ls command will crash with:

UnicodeEncodeError: 'charmap' codec can't encode character '\u2714' in position 0: character maps to <undefined>

Rather than crash, we instead want to make a best effort
attempt to print the key name.  The new behavior is to print
a placeholder character for any character that can't be encoded
into the target encoding.  Example output:

2016-01-12 13:43:34         11 sample_keyname?Foo.txt