Propagate error on S3 range download failures
When we downlaod a file in parts, the DownloadPartTask
was not adding a print task error on the result queue.
As a result, the request would appear to succeed, not print
any error message, and exit with a non zero rc.

I also audited all the ``except Exception`` clauses in tasks.py
and made sure every other place was adding an ``error=True`` task.
DownloadPartTask was the only one missing this.