Remove back-compat shorthand features from new services
The way we currently handle the shorthand parser means that a service
which has an arument that is modeled as a list of structures with a
single member can never add new members. This makes what was a perfectly
fine change into a breaking change. We can't change this behavior for
existing services, but this will prevent the problem from cascading to
new services.