Use a shared bucket across all test_plugin tests
This test adds a setup/teardown at the module level that
creates a bucket that all tests can use.  This change
was made for several reasons:

* Reduce number of resources created.  Considering multiple test runs
  on multiple python version, multiple platforms, etc.
* Have a more robust check in place for creating a bucket.  This
  better handles eventual consistency.
* Improve speed.  Rather than having to create/delete a bucket each
  test, we can reuse a bucket across all tests.  As part of this
  change, the shared bucket is emptied out before each test.
  From my testing, this shaves about 3 minutes off the test runs.

There are some tests that either require a specific bucket (a bucket
in a specific region) or multiple buckets.  This PR does not address
those cases.