Flush stdout after input prompt.
Some terminal emulators (such as mintty) will not write output to the
terminal window until stdout is explicitly flushed. When running `aws
configure` a user would see nothing but a new line. After pressing
'enter' four times, they would then see the output flushed all in a
single line. This makes it very difficult to configure the command.

The solution is to call `flush` after every prompt. Since `raw_input`
does not have an option to do this, we have to prompt and flush
manually. Since we're already accessing stdout directly, it's easier to
write to it directly since print adds some formatting that we don't
want (namely, a newline at the end of the print).

Fixes #1925