Wrap sys.stdin to force non seekable behavior
S3Transfer changes its upload logic based on whether or not
a fileobj is seekable.

On windows, seekable(sys.stdin) was returning True despite the
fact that sys.stdin does not provide correct seekable behavior.

This would result in streaming uploads hanging.

The fix here is to ensure that seekable(fileobj) returns False
for sys.stdin.  To accomplish this, sys.stdin is wrapped in
an object that does not have a seek() method.  This ensures that
S3Transfer will choose the correct code path, and streaming uploads
on windows will not hang.