s3.sync: Do not stat files that don't match filters
The way the task pipeline that each command executes filters files
*after* collecting information about it. For local files it means that
even when the user doesn't want to sync these files, they'll still be
stat'ed, which is unnecessary.

This change uses the filters received from the command line inside of
the `FileGenerator.should_ignore_file` method. So the list of files to
be filtered won't contain files supposed to be excluded.