Change the defaults for ResultPrinter
Instead of using sys.stderr and sys.stdout, make the defaults None. If
you define sys.stderr and sys.stdout as the default, it breaks the functional
tests when sys.stdout and sys.stderr are patched out as the patch happens
after the definition of the initializer and so they do not get patched out.