Switch subscriber additions to use append
Before insert was used because the ResultSubscriber was always provided
at the beginning of the list and additional subscribers needed to be added
in front of it because the ResultSubscriber needs to be last. Now the
ResultSubscriber is added after any custom subscribers have been added to
help ensure this order. Note that the changes in the orders of the test does
not really matter. The suscribers that got moved around do not depend on one
another.