Add support for recording progress of unknown size
This is specifically needed for streaming uploads where the size may not
be known a head of time but may be provided at any point. While it is not
necessarily needed for supporting streaming uploads because we do not print
out progress, it is useful if we ever wanted this statistic for something else
like for the sake of tracking transfer speeds later on.