Fix bug where previous files were not being removed
The benchmark-rm file creates a backup of the files
it's about to delete.  After each iteration of the running
the "aws rm" command, it copies files from the backup location
over to the original location.  This allows multiple iterations
of the script to run.

However, this also means that after running `benchmark-rm`, the
files in s3 still exist and need to be cleaned up once the script
has finished running.  If you don't clean them up
(which is what was happening here), then the subsequent run
will 'rm' both the files from the current
perf run as well as the files from the *previous* perf run.