Improve output progress for s3 commands
Progress would only be outputted if the current number of queued
transfers did not equal the number of completed transfers. This
is an issue for commands like sync because the results queue can
be completely processed for long periods of time causing the
progress to disappear. So instead, print progress as well if
there has not been a notification for the final expected files
to be transferred.