Fix issue when using only show errors printer
The final submission task would at times print out just a single
newline with no padding as the logic before was to clear the progress
bar if the final expected files transferred was not 0 and it was the
last result that the would be processed. Instead the logic should
be if there is even progress displayed in the first place and if
it is the final processed result; there is no reason to be trying
to clear the progress if there is no progress displayed in the
first place. Also note it handles the case of when the final expected
files transferred is 0 because if no transfers happened there will
never have been progress.