Enable local file deletes from fileinfos
This is specifically done for sync --delete downloads. The
implementation is a bit quirky because the functionality is needed
for the CLI but does not really belong in s3transfer. Essentially
what happens if a local delete is seen, the CLI will submit the
transfer request and immediately delete the file in the main thread
as a part of the submission. While it is not ideal to be
doing that in the main thread because it will slow down the process,
it is acceptable as deleting a local file is pretty quick.