Upgrade virtualenv for bundled installers
The upgrade is needed due to the version of wheels pulled
in. When pulling in PyYAML as a dependency, PyYAML tries to
import a function from wheels that throws an IOError on
some patched Ubuntu versions. The issue is similar to:
https://github.com/pypa/pip/issues/1074

Wheels has a fix or that bug and virtualenv
pulls that version of wheels in so it should no longer be an
issue.