Move start time recording to first queued result
Before, the start time would be recorded in the first progress
result. However, this results in less accurate transfer speed
tracking at the beginning of a CLI command as there may be n
number of threads making a request and sending their initial
progress result at about the same time causing the transfer
speed to appear higher then it should be.

As a side note, it was considered to add a timestamp to
QueuedResult, but it was decided against because it seemed
unnecessary if it was only going to be used for the first
received QueuedResult and instead the start time is set
on the first invocation of the handler for QueuedResult