Handle race conditions in s3 sync a little better
The LBYL approach with should_ignore_file checking if triggers_warning
is True is susceptible to rare race conditions in which a file is
deleted after should_ignore_file calls triggers_warning but before
get_file_stat calls os.stat. When that happens, an OSError is raised.
This doesn't affect most people, but when you are using s3 sync on a
large directory that is constantly shifting (for example, a logging
directory being managed by logrotate, or a directory with a ton of tmp
files that come and go) this has a much higher chance of happening.

While my solution here does not address every possible error case, it
does address a few. I'm not sure how I can create a robust blanket
solution, but a little improvement is better than none at all, right?

Fixes #1841