Refresh Bundled Installer
Refresh the bundled installer, moving over to primarily using wheels
which will speed up installation of the installer as well as removes
the need (or even attempt) of a compiler for PyYAML on the end user
machine. In addition, it allows specifying alternative locations for
awscli and botocore, which will typically be used to specify a branch
to install botocore from.