Don't import awscli inside of setup.py
It's a not entirely supported setup to import the package being
installed inside of the setup.py, although it often times can work
if you're careful. However, this breaks down in some edge cases and
it is a better idea to avoid doing that.

This takes the same technique used in pip, and applies it to awscli.