Fix ec2 pagination bug
Resolves #2452

Fixes a bug where cli json input would get processed after ec2
auto pagination injection.

This was due to a subtle behavior of the event system. A handler
registered at the top level without any delimiters will always
be called *after* any events registered that do have delimiters
regarless of when or how they were registered.

The handler for --cli-input-json was registered against
'calling-command', and the handler for the ec2 pagination injection
was registered against 'calling-command.ec2.operation-name'. So then
even though the ec2 pagination injection was using register_last, it
was being called first.

The solution is to use 'calling-command.*' instead for the event
that --cli-input-json is registered against. This puts it in the same
pool as the other delimited handlers. This had the potential to break
other handlers depending on the existing ordering, but there were
very few registered to that event. Running the tests revealed no
issues, however.