Changed waiter.config.delay from 5 seconds to 10 seconds
The documentation below mentions the following with regards to: aws cloudformation wait stack-create-complete [1]
[1] http://docs.aws.amazon.com/cli/latest/reference/cloudformation/wait/stack-create-complete.html

"Wait until stack status is CREATE_COMPLETE. It will poll every 30 seconds until a successful state has been reached. 
This will exit with a return code of 255 after 120 failed checks."

This should be technically be 30 seconds not 5 seconds or 10 seconds because: 
120 calls at 30 seconds intervals will equal 3600 seconds/1 hour until the "Max attempts exceeded" which to confirms with CloudFormations 1 hour* timeout.

I understand that it is a fine line between performance (speed) and consistency, but unfortunately some services like CloudFront take time to build.