Remove mapping of exception from os.stat
This mapping has no effect on Python 2.x, because os.stat raises an
OSError if the file is not found, not an IOError. Python 3.3+ raise a
FileNotFoundError, so the mapping provides no additional context.

It does however throw off the handling of this error later, so this
change fixes #2960.

To make sure the error handling works correctly, throw the correct
exception depending on the python version running the test. Remove the
unit test that tested this conversion; replacing it makes no sense
because we would only be testing the behaviour of the mock.