Don't call describe_stacks in 'aws cloudformation deploy'
We're running into the following throttling errors when trying to call
`aws cloudformation deploy` multiple times both sequentially across
multiple deployment machines and in parallel on a single machine.

```
Waiting for changeset to be created..
An error occurred (Throttling) when calling the DescribeStacks operation (reached max retries: 4): Rate exceeded
Command exited with non-zero status 255
```

We've opened support case 4890801861 in regards to this issue where it
was determined that the `aws cloudformation deploy` command is actually
calling the cloudformation `describe_stacks` command under the hood to
determine if a stack exists or not.

Since we have so many CloudFormation stacks the `describe_stacks` API
can be slow and ineffecient so this pull request replaces those calls
with ones to `describe_stack_resources` instead. It retains the same
behavior for determining if the deploy should be performed with either
a `create` or an `update` command for the following scenarios:

* If the stack does not exist then `create` is used
* If the stack exists in `REVIEW_IN_PROGRESS` state then `create` is used
* Otherwise if the stack already exists then `update` is used

Here are a couple links I found where this issue was previously reported.

* https://aws.uservoice.com/forums/598381-aws-command-line-interface/suggestions/33168517--aws-cloudformation-deploy-wait-describe-stack
* https://github.com/aws/aws-cli/issues/2948