Launch Spot EMR Clusters Without Providing BidPrice
As a result of changes made by the EC2 Spot team, providing a spot price
when requesting a spot instance is now optional. The purpose of the
following code changes is to enable users to launch instance groups
using the AWS EMR ClI without explicitly specifying a bid price.

In this commit, logic in instancegrouputils is modified to set the
MarketType to spot, but not set a BidPrice, when a customer specifies
BidPrice=OnDemandPrice using the --instance-groups option when creating
clusters. Unit tests are added as appropriate.

EMR will handle the null BidPrice value provided by the CLI and launch
spot clusters. This specific implentation was chosen because it was the
easiest way for users to quickly launch spot clusters and did not
require introducing new flags or options in the CLI. It also ensures
users understand that the BidPrice implicitly used will be equal to the
OnDemandPrice.