Defer handling of beta versions
This updates make-bundle slightly to be a tad more lenient in
checking required versions. It will check version numbers in
descending order rather than checking every part. This means that
we are less likely to have to compare beta version parts and other
esoterica. A more robust solution would be to pull in packaging, but
that would introduce dependencies where previously there were none.