Add configuration option to opt-out of url parameter fetching
By default the CLI will fetch the content from any string parameter that
starts with https:// or http:// and use the content as the parameter
rather than the URL itself. This adds a configuration option to block
this behavior and these two cases the same as any other string argument.

Specifically the newly added config option is the `cli_follow_urlparam`
option. It can be used to disable this behavior by adding the following
to the config file in `.aws/config`.

```
cli_follow_urlparam = false
```

If the value is set to `true` or if the config key is not present then the
default behavior is preserved.