Add top level parameters for passing images to rekognition
Raw bytes cannot be passed to nested parameters. Rekognition has many
operations that accept an image as a nested parameter, as essentially a
tagged union between the raw local file, or an S3 location. In the CLI
the S3 location specification works, and the raw image bytes option does
not. This commit adds a new top level parameter that overwrites the
nested image bytes option, which allows for the use of the `fileb://`
prefix to load a file.

For example the detect-labels operation would look like this:

    aws rekognition detect-labels --image-file fileb://image.png

The new top level --image-file parameter fill in the Bytes value of the
existing --image parameter.