Don't rely on specific message for digest error
Python 3.7 changed the kind of error message you get when a base64
string has improper padding. Rather than just telling you the padding is
messed up, it tells you what exactly is messed up about it. This updates
a test we had that was assuming the specific message wouldn't change.