Add subcommand/arg_table properties to clidriver
This provides a consistent interface across built in commands and custom
commands for accessing both the subcommand and argument tables.  The
custom command class already exposed these properties so I just added
them to the builtin command objects.  This allows us to traverse command
structures without having to use the help commands.  The help may be
changing so we want to avoid coupling autocompletion to the help command
interface.